refactor: Break out getIsPlayheadAttachedToFocusRange into own file
