+++ theatre/studio/src/UIRoot/useKeyboardShortcuts.ts
@@ -3,11 +3,9 @@ import getStudio from '@theatre/studio/getStudio'
 import {cmdIsDown} from '@theatre/studio/utils/keyboardUtils'
 import {getSelectedSequence} from '@theatre/studio/selectors'
 import type {$IntentionalAny} from '@theatre/shared/utils/types'
-import type {IDerivation} from '@theatre/dataverse'
-import {Box, prism, val} from '@theatre/dataverse'
+import {prism, val} from '@theatre/dataverse'
 import type {IPlaybackRange} from '@theatre/core/sequences/Sequence'
-import type Sequence from '@theatre/core/sequences/Sequence'
-import memoizeFn from '@theatre/shared/utils/memoizeFn'
+import {getPlaybackStateBox} from './getPlaybackStateBox'
 
 export default function useKeyboardShortcuts() {
   const studio = getStudio()
@@ -149,45 +147,3 @@ export default function useKeyboardShortcuts() {
     }
   }, [])
 }
-
-type ControlledPlaybackStateBox = Box<
-  undefined | IDerivation<{range: IPlaybackRange; isFollowingARange: boolean}>
->
-
-const getPlaybackStateBox = memoizeFn(
-  (sequence: Sequence): ControlledPlaybackStateBox => {
-    const box = new Box(undefined) as ControlledPlaybackStateBox
-    return box
-  },
-)
-
-/*
- * A memoized function that returns a derivation with a boolean value.
- * This value is set to `true` if:
- * 1. the playback is playing and using the focus range instead of the whole sequence
- * 2. the playback is stopped, but would use the focus range if it were started.
- */
-export const getIsPlayheadAttachedToFocusRange = memoizeFn(
-  (sequence: Sequence) =>
-    prism<boolean>(() => {
-      const controlledPlaybackState =
-        getPlaybackStateBox(sequence).derivation.getValue()
-      if (controlledPlaybackState) {
-        return controlledPlaybackState.getValue().isFollowingARange
-      } else {
-        const {projectId, sheetId} = sequence.address
-        const focusRange = val(
-          getStudio().atomP.ahistoric.projects.stateByProjectId[projectId]
-            .stateBySheetId[sheetId].sequence.focusRange,
-        )
-
-        if (!focusRange || !focusRange.enabled) return false
-
-        const pos = val(sequence.pointer.position)
-
-        const withinRange =
-          pos >= focusRange.range.start && pos <= focusRange.range.end
-        return withinRange
-      }
-    }),
-)
